requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
requests
openai
PyMuPDF
pdf2image
pytesseract
tiktoken
httpx
python-multipart
passlib
python-jose
numpy==1.24.4
scikit-learn
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
git+https://github.com/openai/whisper.git
faiss-cpu==1.7.4
aiohttp
aiofiles